move project view files around
