+++ cory/src/components/Clip.jsx
